fixed merge conflict issues
